Add a filtercheck for is_different_addr
In order to do some UDP tests, we want to know when a given
sendto/recvfrom on a socket was for a "new" source or destination. So
add a filtercheck fd.is_different_addr for that. It is initially false,
becomes true when a sendto/recvfrom changes the source or destination,
and becomes false on a second sendto/recvfrom with the same
source/destination.

It's implemented by relying on the return value from update_fd(), which
returns true when the source/destination changes.